requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Pillow==10.1.0
Werkzeug==3.0.1